requirements.txt
opencv-python>=4.8.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
numpy>=1.23.0,<2.0.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
Pillow>=9.5.0
scipy>=1.10.0
pandas>=1.5.0
seaborn>=0.12.0
jupyter>=1.0.0
jupyterlab>=4.0.0